requirements.txt
sendgrid==6.7.0
discord.py==1.7.1
asyncio==3.4.3

